+++ bot/handlers/callbacks.py
@@ -1,11 +1,11 @@
 """
 /bot/handlers/callbacks.py
-Server Monitoring System v8.32.45
+Server Monitoring System v8.32.46
 Copyright (c) 2025 Aleksandr Sukhanov
 License: MIT
 A single router for callbacks.
 Система мониторинга серверов
-Версия: 8.32.45
+Версия: 8.32.46
 Автор: Александр Суханов (c)
 Лицензия: MIT
 Единый router callback’ов.