+++ config/db_settings_app.py
@@ -1,11 +1,11 @@
 """
 /config/db_settings_app.py
-Server Monitoring System v8.32.45
+Server Monitoring System v8.32.46
 Copyright (c) 2025 Aleksandr Sukhanov
 License: MIT
 Database Settings Manager
 Система мониторинга серверов
-Версия: 8.32.45
+Версия: 8.32.46
 Автор: Александр Суханов (c)
 Лицензия: MIT
 Менеджер настроек БД